udate game2048 + main-refactor added move methods
