+++ src/scripts/main-refactor.js
@@ -2,8 +2,8 @@ import { Game } from './game2048.js';
 import { isArrowButton } from './utils.js';
 
 const game = new Game();
-game.createChip();
-console.log(game.cellArr, game.arrChip);
+// game.createChip();
+// console.log(game.cellArr, game.arrChip);
 
 document.addEventListener('keydown', (ev) => {
   const direction = ev.key; // direction arrow.
@@ -12,7 +12,20 @@ document.addEventListener('keydown', (ev) => {
     return;
   }
 
-  game.moveLeft(direction);
+  if (direction === 'ArrowUp') {
+    game.moveUp(direction);
+  }
+  if (direction === 'ArrowDown') {
+    game.moveDown(direction);
+  }
+  if (direction === 'ArrowLeft') {
+    game.moveLeft(direction);
+  }
+  if (direction === 'ArrowRight') {
+    game.moveRight(direction);
+  }
 
+  game.resetMergeFlags();
   game.createChip(); // OBJ Chip
+  game.renderHtmlChip();
 });
